feat: added console.log for debugging
in http://127.0.1.1:5000/test/game.html, sending a message will trigger a getGameData event. Check the console for a list of the current state of the game. This is useful when you use http://127.0.1.1:5000/gm/dummygame to set the round status.

BREAKING CHANGE: set line 10 in main.py to your own logic file for debugging.
BREAKING CHANGE: for the GMs screen, the output of the "update" Button has moved to the JS console. Hit Ctrl-Shift-C to open the console.